[Refactor] Modularized form components
