+++ frontend/src/pages/Login.jsx
@@ -1,9 +1,7 @@
-import Form from "../components/Form";
+import Form from "../components/Forms/Form";
 
 const Login = () => {
-    return (
-        <Form route="/api/token/" method="login"/>
-    )
-}
+  return <Form route="/api/token/" method="login" />;
+};
 
 export default Login;
