+++ frontend/src/pages/Register.jsx
@@ -1,7 +1,7 @@
-import Form from "../components/Form";
+import Form from "../components/Forms/Form";
 
 const Register = () => {
-    return <Form route="/api/user/register/" method="register" />
-}
+  return <Form route="/api/user/register/" method="register" />;
+};
 
 export default Register;
